Made changes
Synchro requested me to make the delete thing toggleable, so here we are

+++ main.py
@@ -1,16 +1,24 @@
 import discord
 import json
+from os import path
 
 class MyClient(discord.Client):
     async def on_ready(self):
       #This basically just says who it is
       print('Logged on as', self.user)
+      with open("urmumgaylulz.txt", "w+") as f:
+        f.write("Uoukprhtp aqmul fu, B aqmu uoukprhtp")
 
     async def on_message(self, message):
-      #Everyone with the name Alfreaca gets ignored
-      #Haha fuck you Alf, you're being skippedd
-      #Nah man jk love you (nohomo tho)
       member = discord.utils.get(message.guild.members, name='Alfreaca')
+
+      if message.content == "||on":
+        with open("urmumgaylulz.txt", "w+") as f:
+          f.write("Uoukprhtp aqmul fu, B aqmu uoukprhtp")
+
+      
+      if message.content == "||off":
+        os.remove("urmumgaylulz.txt")
 
       #I came up with this genius loop so it doesn't loop over it's own messages.
       if member != message.author:
@@ -27,13 +35,14 @@ class MyClient(discord.Client):
         #This is the loop for checking the words.json file
         for data in datas:
           if data in str(message.content).lower():
-            #This loads up the OffenseWarnings channe;
+            #This loads up the OffenseWarnings channel
             modroom = client.get_channel(568732658411110410)
             print(message.author)
 
             #Again, absolutely genius
             if message.author != self.user:
-              await message.delete()
+              if path.exists("urmumgaylulz.txt"):
+                await message.delete()
             
             #Tells momma anyway
             #brothers (◔_◔) amirite
@@ -41,12 +50,10 @@ class MyClient(discord.Client):
 
             #This DM's the dude/dudina to not do it again or it'll tell mommy
             channel = await message.author.create_dm()
-            await channel.send("You said `{}`, which consists of a blacklisted word. You have been added to the list of warnings, meaning that if you do this again, you will be banned/kicked/muted.".format(message.content))
+            await channel.send("You said `{}`, which consists of a blacklisted word. You have been added to the list of warnings, meaning that if you do this again, you will be banned/kicked/muted.".format(message.content))  
 
 #This just runs the class
 client = MyClient()
 
 #This runs the bot
 client.run("#")
-#I ain't falling for it again haha fuck you sync
-#no jk man, love you 2 (nohomo tho)
